requirements.txt
albumentations==1.3.0
matplotlib==3.5.3
numpy==1.21.5
opencv_python_headless==4.7.0.68
Pillow==9.4.0
scikit_learn==1.2.0
scipy==1.7.3
tifffile==2021.11.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
